refactor Faker class
