+++ src/test/java/com/butbetter/applicationservices/faker/Faker.java
@@ -1,0 +1,14 @@
+package com.butbetter.applicationservices.faker;
+
+public class Faker {
+
+    private final AlcoholFaker alcoholFaker;
+
+    public Faker() {
+        this.alcoholFaker = new AlcoholFaker();
+    }
+
+    public AlcoholFaker getAlcohol() {
+        return alcoholFaker;
+    }
+}
