+++ src/test/java/com/butbetter/applicationservices/productapi/service/AlcoholValidationServiceTest.java
@@ -1,8 +1,7 @@
-package com.butbetter.applicationservices.db.service;
+package com.butbetter.applicationservices.productapi.service;
 
-import com.butbetter.applicationservices.Faker.Faker;
+import com.butbetter.applicationservices.faker.Faker;
 import com.butbetter.applicationservices.productapi.model.Alcohol;
-import com.butbetter.applicationservices.productapi.service.AlcoholValidationService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -19,83 +18,83 @@ class AlcoholValidationServiceTest {
 
     @Test
     void checkADefaultProduct() {
-        assertDoesNotThrow((() -> this.alcoholValidationService.checkProduct(faker.getProductFaker().Corona)));
+        assertDoesNotThrow((() -> this.alcoholValidationService.checkProduct(faker.getAlcohol().Corona)));
     }
 
     @Test
     void checkNameIsEmpty() {
         assertThrows(IllegalArgumentException.class, () -> {
-            this.alcoholValidationService.checkProduct(faker.getWrongProduct().emptyName);
+            this.alcoholValidationService.checkProduct(faker.getAlcohol().emptyName);
         });
     }
 
     @Test
     void isNegativePrice() {
         assertThrows(IllegalArgumentException.class, () -> {
-            this.alcoholValidationService.checkProduct(faker.getWrongProduct().negativePrice);
+            this.alcoholValidationService.checkProduct(faker.getAlcohol().negativePrice);
         });
     }
 
     @Test
     void isNegativePercentage() {
         assertThrows(IllegalArgumentException.class, () -> {
-            this.alcoholValidationService.checkProduct(faker.getWrongProduct().negativePercentage);
+            this.alcoholValidationService.checkProduct(faker.getAlcohol().negativePercentage);
         });
     }
 
     @Test
     void isPercentageOver100() {
         assertThrows(IllegalArgumentException.class, () -> {
-            this.alcoholValidationService.checkProduct(faker.getWrongProduct().over100Percentage);
+            this.alcoholValidationService.checkProduct(faker.getAlcohol().over100Percentage);
         });
     }
 
     @Test
     void isNegativeAmount() {
         assertThrows(IllegalArgumentException.class, () -> {
-            this.alcoholValidationService.checkProduct(faker.getWrongProduct().negativeAmount);
+            this.alcoholValidationService.checkProduct(faker.getAlcohol().negativeAmount);
         });
     }
 
     @Test
     void isMinAmount() {
         assertDoesNotThrow(() -> {
-            this.alcoholValidationService.checkProduct(faker.getWrongProduct().minAmount);
+            this.alcoholValidationService.checkProduct(faker.getAlcohol().minAmount);
         });
     }
 
     @Test
     void isMaxAmount() {
         assertThrows(IllegalArgumentException.class, () -> {
-            this.alcoholValidationService.checkProduct(faker.getWrongProduct().maxAmount);
+            this.alcoholValidationService.checkProduct(faker.getAlcohol().maxAmount);
         });
     }
 
     @Test
     void isNegativeAge() {
         assertThrows(IllegalArgumentException.class, () -> {
-            this.alcoholValidationService.checkProduct(faker.getWrongProduct().negativeAgeOfRestriction);
+            this.alcoholValidationService.checkProduct(faker.getAlcohol().negativeAgeOfRestriction);
         });
     }
 
     @Test
     void checkMinAge() {
         assertDoesNotThrow(() -> {
-            this.alcoholValidationService.checkProduct(faker.getWrongProduct().maxAgeOfRestriction);
+            this.alcoholValidationService.checkProduct(faker.getAlcohol().maxAgeOfRestriction);
         });
     }
 
     @Test
     void checkMaxAge() {
         assertDoesNotThrow(() -> {
-            this.alcoholValidationService.checkProduct(faker.getWrongProduct().maxAgeOfRestriction);
+            this.alcoholValidationService.checkProduct(faker.getAlcohol().maxAgeOfRestriction);
         });
     }
 
     @Test
     void checkOverMaxAge() {
         assertThrows(IllegalArgumentException.class, () -> {
-            final Alcohol alcohol = faker.getWrongProduct().maxAgeOfRestriction;
+            final Alcohol alcohol = faker.getAlcohol().maxAgeOfRestriction;
             alcohol.setAgeOfRestrictions(alcohol.getAgeOfRestrictions() + 1);
             this.alcoholValidationService.checkProduct(alcohol);
         });
